+++ src/components/AddressSearch/styles.ts
@@ -1,0 +1,47 @@
+import * as stylex from "@stylexjs/stylex";
+
+export const styles = stylex.create({
+  modalOverlay: {
+    position: "fixed",
+    top: "0",
+    left: "0",
+    right: "0",
+    bottom: "0",
+    backgroundColor: "rgba(0, 0, 0, 0.5)",
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+    zIndex: "50",
+  },
+  modalContent: {
+    backgroundColor: "#fff",
+    borderRadius: "12px",
+    width: "90%",
+    maxWidth: "500px",
+    maxHeight: "90vh",
+    overflow: "hidden",
+  },
+  modalHeader: {
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "space-between",
+    padding: "16px",
+    borderBottom: "1px solid #DDE1E6",
+  },
+  modalTitle: {
+    fontSize: "18px",
+    fontWeight: "600",
+  },
+  closeButton: {
+    backgroundColor: "transparent",
+    border: "none",
+    fontSize: "20px",
+    cursor: "pointer",
+    padding: "4px",
+    color: "#666",
+  },
+  postcodeContainer: {
+    width: "100%",
+    height: "400px",
+  },
+});
